third_party/WebKit/LayoutTests/web-animations-api/player-finish-sample-only-expected.txt

PASS: playerTop is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
